lints & split big deals module
